Local (aka UNIX) socket transport support.
This is not a popular transport for SNMP, but FreeBSD's snmp daemon
supports it.  The only official mentioning of local socket for SNMP can be
found in RFC 3419, but it doesn't differentiate between stream and
datagram socket.

The UNIX::Stream inherits from IPv4::TCP as it needs reassembly, and
UNIX::Datagram inherits from IPv4::UDP.  The UNIX::Datagram requires
bind(2)-ing, cause otherwise the SNMP daemon is not able to reply.  This
is a known limitation of datagram UNIX sockets.  The UNIX::Stream skips
binding.